removing lib/events
