+++ lib/events.js
@@ -1,91 +1,82 @@
-/**
- * @module lib/events
- */
-'use strict';
-require('loadenv')({ debugName: 'docker-listener' });
-
-var ip = require('ip');
-var os = require('os');
-var uuid = require('node-uuid');
-var ErrorCat = require('error-cat');
-var errorCat = new ErrorCat();
-
-var docker = require('./docker');
-var log = require('./logger').getChild(__filename);
-
-// added here because we only want to call these once for speed
-var numCpus = os.cpus().length;
-var tags = process.env.HOST_TAGS + '';
-var hostIp = ip.address();
-var totalmem = os.totalmem();
-
-/**
- * adds extra data to events
- * @param {Object} dockerEvent
- * @param {Function} cb
- */
-exports.enhance = function (dockerEvent, cb) {
-  dockerEvent = addBasicFields(dockerEvent);
-  if (!isContainerEvent(dockerEvent)) {
-    return cb(null, dockerEvent);
-  }
-  addInspect(dockerEvent, function(err, ev) {
-    // ignore this error, we want to send container deaths with or without data
-    // but we still want to log it
-    if (err) {
-      errorCat.log(err);
-      ev = dockerEvent;
-    }
-    cb(null, ev);
-  });
-};
-
-/**
- * add `ip`, `uuid`, `host` and `time` (if missing) fields
- * @param {Object} dockerEvent
- * @return Object
- */
-function addBasicFields (dockerEvent) {
-  dockerEvent.uuid = uuid.v1();
-  dockerEvent.ip = hostIp;
-  dockerEvent.numCpus = numCpus;
-  dockerEvent.mem = totalmem;
-  dockerEvent.tags = tags;
-  if (!dockerEvent.time) {
-    dockerEvent.time = new Date().getTime();
-  }
-  dockerEvent.host = 'http://' + dockerEvent.ip + ':' + process.env.DOCKER_REMOTE_API_PORT;
-  return dockerEvent;
-}
-
-/**
- * should add inspect data if dockerEvent is a container change dockerEvent
- * should only retry a few times
- */
-function addInspect (dockerEvent, cb) {
-  var container = docker.getContainer(dockerEvent.id);
-  container.inspect(function (err, data) {
-    if (err) {
-      log.error({
-        err: err,
-        dockerEvent: dockerEvent
-      }, 'container inspect error');
-      return cb(err);
-    }
-    log.info({
-      dockerEvent: dockerEvent,
-      data: data
-    }, 'container inspect complete');
-    dockerEvent.inspectData = data;
-    cb(null, dockerEvent);
-  });
-}
-
-/**
- * should return true if the dockerEvent is a container dockerEvent
- */
-function isContainerEvent (dockerEvent) {
-  var containerEvent =
-    ['create', 'die', 'export', 'kill', 'pause', 'restart', 'start', 'stop', 'unpause'];
-  return ~containerEvent.indexOf(dockerEvent.status);
-}
+// /**
+//  * @module lib/events
+//  */
+// 'use strict';
+// require('loadenv')({ debugName: 'docker-listener' });
+//
+// var ip = require('ip');
+// var os = require('os');
+// var uuid = require('node-uuid');
+// var ErrorCat = require('error-cat');
+// var errorCat = new ErrorCat();
+//
+// var docker = require('./docker');
+// var log = require('./logger').getChild(__filename);
+//
+// // added here because we only want to call these once for speed
+// var numCpus = os.cpus().length;
+// var tags = process.env.HOST_TAGS + '';
+// var hostIp = ip.address();
+// var totalmem = os.totalmem();
+//
+// /**
+//  * adds extra data to events
+//  * @param {Object} dockerEvent
+//  * @param {Function} cb
+//  */
+// exports.enhance = function (dockerEvent, cb) {
+//   dockerEvent = addBasicFields(dockerEvent);
+//   if (!isContainerEvent(dockerEvent)) {
+//     return cb(null, dockerEvent);
+//   }
+//   addInspect(dockerEvent, function(err, ev) {
+//     // ignore this error, we want to send container deaths with or without data
+//     // but we still want to log it
+//     if (err) {
+//       errorCat.log(err);
+//       ev = dockerEvent;
+//     }
+//     cb(null, ev);
+//   });
+// };
+//
+// /**
+//  * add `ip`, `uuid`, `host` and `time` (if missing) fields
+//  * @param {Object} dockerEvent
+//  * @return Object
+//  */
+// function addBasicFields (dockerEvent) {
+//   dockerEvent.uuid = uuid.v1();
+//   dockerEvent.ip = hostIp;
+//   dockerEvent.numCpus = numCpus;
+//   dockerEvent.mem = totalmem;
+//   dockerEvent.tags = tags;
+//   if (!dockerEvent.time) {
+//     dockerEvent.time = new Date().getTime();
+//   }
+//   dockerEvent.host = 'http://' + dockerEvent.ip + ':' + process.env.DOCKER_REMOTE_API_PORT;
+//   return dockerEvent;
+// }
+//
+// /**
+//  * should add inspect data if dockerEvent is a container change dockerEvent
+//  * should only retry a few times
+//  */
+// function addInspect (dockerEvent, cb) {
+//   var container = docker.getContainer(dockerEvent.id);
+//   container.inspect(function (err, data) {
+//     if (err) {
+//       log.error({
+//         err: err,
+//         dockerEvent: dockerEvent
+//       }, 'container inspect error');
+//       return cb(err);
+//     }
+//     log.info({
+//       dockerEvent: dockerEvent,
+//       data: data
+//     }, 'container inspect complete');
+//     dockerEvent.inspectData = data;
+//     cb(null, dockerEvent);
+//   });
+// }
