+++ lib/workers/docker.event.publish.js
@@ -7,12 +7,23 @@
 require('loadenv')()
 
 var ip = require('ip');
+var os = require('os');
+var uuid = require('node-uuid');
+var ErrorCat = require('error-cat');
+var errorCat = new ErrorCat();
 var keypather = require('keypather')()
+var docker = require('../docker');
 
 var events = require('../events')
 var rabbitmq = require('../rabbitmq')
 var Promise = require('bluebird')
 var log = require('../logger').getChild(__filename)
+
+// added here because we only want to call these once for speed
+var numCpus = os.cpus().length;
+var tags = process.env.HOST_TAGS + '';
+var hostIp = ip.address();
+var totalmem = os.totalmem();
 
 module.exports = DockerEventPublish
 
@@ -21,9 +32,19 @@ function DockerEventPublish (job) {
     tx: true,
     data: job
   }
-  var data = job.data
+  var data = addBasicFields(job.data)
   return Promise.fromCallback(function (cb) {
-      events.enhance(data, cb)
+      if (!isContainerEvent(data)) {
+        return cb(null, data)
+      }
+      var container = docker.getContainer(dockerEvent.id);
+      container.inspect(function (err, inspectData) {
+        if (err) {
+          return cb(err)
+        }
+        data.inspectData = inspectData
+        cb(null, data)
+      })
     })
     .then(function (enhanced) {
       /**
@@ -77,6 +98,34 @@ function isBuildContainer (enhanced) {
   return keypather.get(enhanced, 'inspectData.Config.Labels.type') === 'image-builder-container';
 }
 
+
+/**
+ * should return true if the dockerEvent is a container dockerEvent
+ */
+function isContainerEvent (dockerEvent) {
+  var containerEvent =
+    ['create', 'die', 'export', 'kill', 'pause', 'restart', 'start', 'stop', 'unpause'];
+  return ~containerEvent.indexOf(dockerEvent.status);
+}
+
+/**
+ * add `ip`, `uuid`, `host` and `time` (if missing) fields
+ * @param {Object} dockerEvent
+ * @return Object
+ */
+function addBasicFields (dockerEvent) {
+  dockerEvent.uuid = uuid.v1();
+  dockerEvent.ip = hostIp;
+  dockerEvent.numCpus = numCpus;
+  dockerEvent.mem = totalmem;
+  dockerEvent.tags = tags;
+  if (!dockerEvent.time) {
+    dockerEvent.time = new Date().getTime();
+  }
+  dockerEvent.host = 'http://' + dockerEvent.ip + ':' + process.env.DOCKER_REMOTE_API_PORT;
+  return dockerEvent;
+}
+
 /**
  * returns routing key. format:
  * <org>.<host_ip>
@@ -87,6 +136,6 @@ function isBuildContainer (enhanced) {
 DockerEventPublish._createRoutingKey = function () {
   return [
     process.env.HOST_TAGS.split(',')[0],
-    ip.address().replace('.', '-')
+    hostIp.replace('.', '-')
   ].join('.')
 }
